requirements.txt
baker==1.3
python-binance==0.7.9
numpy==1.19.5
pandas==1.2.0
psycopg2-binary==2.8.6
sqlalchemy==1.3.22